Started  work on Task 5.
Stuck on "TestToStringStartsAndEndsWithNewLine" not sure how to approach the problem. I used Environment.NewLine to build the string returned by ToString() . Would \n be better to allow split string method in this test? Should I explore StringReader or StringWriter? All Stack Overflow discussions I have found seem to show how to access specific lines in separate files, not in local string variables.

Two TestTask5 tests are having trouble locating files. These may be the broken tests?

Unsure why line 102 in TestTask5 is returning false as the contains statements in that line should be true as the text they are checking for appears to be in the string returned by ToString().